backend/requirements.txt
# Django Core
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2

# Authentication
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Cache & Task Queue
redis==5.0.4
celery==5.3.6
django-redis==5.4.0

# File Storage & Media
Pillow>=10.4.0
python-magic==0.4.27

# PDF Generation
WeasyPrint>=62.1
reportlab==4.2.0

# Environment Variables
python-decouple==3.8
python-dotenv==1.0.1

# WSGI Server
gunicorn==22.0.0

# Utilities
pytz==2024.1
python-dateutil==2.9.0

# Development
django-extensions==3.2.3

# Testing
pytest==8.2.0
pytest-django==4.8.0
faker==25.0.0

# Code Quality
black==24.4.2
